64비트 대응 (포인터 크기가 커짐으로인한 이슈 해결)
git-svn-id: svn://www.hshgames.com/project_s@29 e3c12c2a-3f9b-4cee-bc1f-3a9786f174a7

+++ proj.win32/Classes/Task/PoolingManager.h
@@ -66,7 +66,7 @@ private:
 	std::vector<MEMORYBLOCK> m_ShooterList;
 
 	/* 메모리 블럭 하나 당 크기 */
-	int m_BulletSize;
-	int m_ShooterSize;
+	size_t m_BulletSize;
+	size_t m_ShooterSize;
 };
